requirements.txt
flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1
pymysql==1.1.1
numpy==2.1.1
